Moved whole universal code from  CurrentMapViewModel to FilteredPOIViewBase. Added filtering and Internet/GPS connection checking to AugmentedViewModel.
