+++ AugmentedSzczecin/AugmentedSzczecin/Services/HttpService.cs
@@ -35,17 +35,17 @@ namespace AugmentedSzczecin.Services
             HttpResponseMessage response = null;
             if (category == CategoryType.ALL)
             {
-                response = await _client.GetAsync(string.Format("q?lt={0}&lg={1}&radius={2}", latitude, longitude, radius));
+                //response = await _client.GetAsync(string.Format("q?lt={0}&lg={1}&radius={2}", latitude, longitude, radius));
             }
             else
             {
-                response = await _client.GetAsync(string.Format("q?lt={0}&lg={1}&radius={2}&cat={3}", latitude, longitude, radius, category));
+                //response = await _client.GetAsync(string.Format("q?lt={0}&lg={1}&radius={2}&cat={3}", latitude, longitude, radius, category));
             }
-            response.EnsureSuccessStatusCode();
-            string json = await response.Content.ReadAsStringAsync();
+            //response.EnsureSuccessStatusCode();
+            //string json = await response.Content.ReadAsStringAsync();
 
             /****** DO TESTÓW ******/
-            //string json = "{\"places\":[{\"name\":\"Test\", \"description\":\"Opis testowy\",\"location\":{\"latitude\":53.4884,\"longitude\":14.5581},\"address\":{\"city\":\"Szczecin\",\"street\":\"Szczecin\",\"streetNumber\":\"11\",\"zipcode\":\"12-321\"},\"tags\":[\"tag\"],\"www\":\"www.blstream.com\",\"phone\":\"123654789\",\"fanpage\":\"fb.com\",\"opening\":[{\"day\":\"MONDAY\",\"open\":\"6:00\",\"close\":\"20:00\"},{\"day\":\"WEDNESDAY\",\"open\":\"10:00\",\"close\":\"18:00\"},{\"day\":\"SATURDAY\",\"open\":\"10:00\",\"close\":\"19:00\"}],\"category\":\"PLACE\",\"subcategory\":\"PARK\"}, {\"name\":\"osmy Budynek pierwszego\",\"location\":{\"longitude\":14.5580,\"latitude\":53.4299},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"PLACE\"}, {\"name\":\"piaty Budynek pierwszego\",\"location\":{\"longitude\":14.5590,\"latitude\":53.4390},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"PLACE\"}], \"events\":[{\"name\":\"Cafe Barbakan\",\"description\":\"Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy\",\"location\":{\"longitude\":14.5588,\"latitude\":53.4299},\"address\":{\"city\":\"Szczecin\",\"street\":\"Szczecinska\",\"streetNumber\":\"11\",\"zipcode\":\"12-321\"},\"tags\":[\"tag1\",\"tag2\",\"tag3\"],\"www\":\"www.blstream.com\",\"phone\":\"123654789\",\"fanpage\":\"fb.com\",\"opening\":[{\"day\":\"MONDAY\",\"open\":\"6:00\",\"close\":\"20:00\"},{\"day\":\"WEDNESDAY\",\"open\":\"10:00\",\"close\":\"18:00\"},{\"day\":\"SATURDAY\",\"open\":\"10:00\",\"close\":\"19:00\"}],\"category\":\"PLACE\",\"subcategory\":\"PARK\"}, {\"name\":\"czwarty Budynek pierwszego\",\"location\":{\"longitude\":14.5590,\"latitude\":53.4390},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"EVENT\"}], \"commercial\":[{\"name\":\"dziewiaty Budynek pierwszego\",\"location\":{\"longitude\":14.5570,\"latitude\":53.4300},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"COMMERCIAL\"}]}";
+            string json = "{\"places\":[{\"name\":\"Punkt 1\", \"description\":\"Opis testowy\",\"location\":{\"latitude\":53.397271,\"longitude\":14.526047},\"category\":\"PLACE\"},{\"name\":\"Punkt 2\", \"description\":\"Opis testowy\",\"location\":{\"latitude\":53.396501,\"longitude\":14.524154},\"category\":\"EVENT\"},{\"name\":\"Test\", \"description\":\"Opis testowy\",\"location\":{\"latitude\":53.4884,\"longitude\":14.5581},\"address\":{\"city\":\"Szczecin\",\"street\":\"Szczecin\",\"streetNumber\":\"11\",\"zipcode\":\"12-321\"},\"tags\":[\"tag\"],\"www\":\"www.blstream.com\",\"phone\":\"123654789\",\"fanpage\":\"fb.com\",\"opening\":[{\"day\":\"MONDAY\",\"open\":\"6:00\",\"close\":\"20:00\"},{\"day\":\"WEDNESDAY\",\"open\":\"10:00\",\"close\":\"18:00\"},{\"day\":\"SATURDAY\",\"open\":\"10:00\",\"close\":\"19:00\"}],\"category\":\"PLACE\",\"subcategory\":\"PARK\"}, {\"name\":\"osmy Budynek pierwszego\",\"location\":{\"longitude\":14.5580,\"latitude\":53.4299},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"PLACE\"}, {\"name\":\"piaty Budynek pierwszego\",\"location\":{\"longitude\":14.5590,\"latitude\":53.4390},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"PLACE\"}], \"events\":[{\"name\":\"Cafe Barbakan\",\"description\":\"Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy Opis testowy\",\"location\":{\"longitude\":14.5588,\"latitude\":53.4299},\"address\":{\"city\":\"Szczecin\",\"street\":\"Szczecinska\",\"streetNumber\":\"11\",\"zipcode\":\"12-321\"},\"tags\":[\"tag1\",\"tag2\",\"tag3\"],\"www\":\"www.blstream.com\",\"phone\":\"123654789\",\"fanpage\":\"fb.com\",\"opening\":[{\"day\":\"MONDAY\",\"open\":\"6:00\",\"close\":\"20:00\"},{\"day\":\"WEDNESDAY\",\"open\":\"10:00\",\"close\":\"18:00\"},{\"day\":\"SATURDAY\",\"open\":\"10:00\",\"close\":\"19:00\"}],\"category\":\"PLACE\",\"subcategory\":\"PARK\"}, {\"name\":\"czwarty Budynek pierwszego\",\"location\":{\"longitude\":14.5590,\"latitude\":53.4390},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"EVENT\"}], \"commercial\":[{\"name\":\"dziewiaty Budynek pierwszego\",\"location\":{\"longitude\":14.5570,\"latitude\":53.4300},\"address\":{\"street\":\"GĹowackiego 3 w dzielnicy Salwator\"},\"subcategory\":\"MONUMENT\",\"category\":\"COMMERCIAL\"}]}";
             /***********************/
 
             if (json == "{}")
