+++ AugmentedSzczecin/AugmentedSzczecin/Views/AugmentedView.xaml
@@ -6,14 +6,98 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:caliburn="using:Caliburn.Micro"
+    xmlns:UserControls="using:AugmentedSzczecin.UserControls"
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
     <Grid>
-        <CaptureElement x:Name="PreviewScreen" Stretch="UniformToFill" />
-        <Canvas x:Name="ItemCanvas" />
-        <Slider x:Name="RadiusSlider" Orientation="Vertical" Minimum="100" Maximum="1000" LargeChange="100" SmallChange="100" Value="{Binding Radius, Mode=TwoWay}" HorizontalAlignment="Right" Margin="0,10" TickFrequency="100" StepFrequency="100" TickPlacement="Inline"/>
-        <TextBlock x:Name="PoiCount" Margin="10" VerticalAlignment="Bottom" HorizontalAlignment="Left" FontSize="32"/>
-        <TextBlock Text="{Binding ElementName=RadiusSlider, Path=Value}" FontSize="32" HorizontalAlignment="Right" VerticalAlignment="Top" Margin="0,10,30,0"/>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="0.5*" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="0.08*" />
+        </Grid.RowDefinitions>
+        <CaptureElement 
+            Grid.Row="0"
+            Grid.RowSpan="3" 
+            x:Name="PreviewScreen" 
+            Stretch="UniformToFill" />
+        <Canvas 
+            Grid.Row="0"
+            Grid.RowSpan="3"
+            x:Name="ItemCanvas" />
+        <Slider 
+            Grid.Row="0"
+            Grid.RowSpan="3"
+            x:Name="RadiusSlider" 
+            Orientation="Vertical" 
+            Minimum="100" 
+            Maximum="1000" 
+            LargeChange="100" 
+            SmallChange="100" 
+            Value="{Binding Radius, Mode=TwoWay}" 
+            HorizontalAlignment="Right" 
+            Margin="0,10" 
+            TickFrequency="100" 
+            StepFrequency="100" 
+            TickPlacement="Inline"/>
+        <TextBlock 
+            Grid.Row="1"
+            x:Name="PoiCount" 
+            Margin="10" 
+            VerticalAlignment="Bottom" 
+            HorizontalAlignment="Left" 
+            FontSize="32"/>
+        <TextBlock 
+            Text="{Binding ElementName=RadiusSlider, Path=Value}" 
+            FontSize="32" 
+            HorizontalAlignment="Right" 
+            VerticalAlignment="Top" 
+            Margin="0,10,30,0"/>
+        <Grid Grid.RowSpan="3" Visibility="{Binding IsFilterPanelVisible, Converter={StaticResource BoolToVisConv}}" >
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="0.2*"/>
+                <ColumnDefinition Width="2*"/>
+                <ColumnDefinition Width="0.2*"/>
+                <ColumnDefinition Width="*"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="*" />
+                <RowDefinition Height="5*" />
+            </Grid.RowDefinitions>
+            <Border 
+                Grid.Column="0" 
+                Grid.ColumnSpan="3" 
+                Grid.RowSpan="2" 
+                Background="{StaticResource PhoneChromeBrush}" />
+            <Border 
+                Grid.Column="4" 
+                Grid.RowSpan="2">
+                <Border.Background>
+                    <LinearGradientBrush EndPoint="0,0.5" StartPoint="1,0.5">
+                        <GradientStop Color="#000" Offset="1"/>
+                        <GradientStop Color="Transparent" Offset="0.0"/>
+                    </LinearGradientBrush>
+                </Border.Background>
+            </Border>
+            <UserControls:SideMenuFilter
+                Grid.Column="1" 
+                Grid.Row="1" 
+                ItemsSourceCategories="{Binding ListOfCategories}" 
+                SelectedCategory="{Binding SelectedValue, Mode=TwoWay}" />
+        </Grid>
     </Grid>
+
+    <Page.BottomAppBar>
+        <CommandBar 
+            IsSticky="True" 
+            x:Name="appBar">
+            <CommandBar.PrimaryCommands>
+                <AppBarButton 
+                    x:Name="ToggleFilter" 
+                    Icon="Filter" 
+                    IsCompact="False" 
+                    Label="Filtrowanie"/>
+            </CommandBar.PrimaryCommands>
+        </CommandBar>
+    </Page.BottomAppBar>
 </Page>
